+++ src/main/java/ru/practicum/shareit/item/Item.java
@@ -3,12 +3,9 @@ package ru.practicum.shareit.item;
 import lombok.Data;
 import ru.practicum.shareit.user.User;
 
-/**
- * TODO Sprint add-controllers.
- */
 @Data
 public class Item {
-    Integer id;
+    Long id;
     String name;
     String description;
     Boolean available;
